ike: Add an option to trigger a DPD instead of a NAT keepalive
This is useful on Android where the app might not be able to send
keep-alives if the device is a sleep for a while.  If the NAT mapping
has been deleted in the mean time, the NAT-D payloads allow detecting
this and doing a MOBIKE update, or recreating the SA if the peer deleted
it because the client wasn't reachable.